requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiogram==2.11.2
aiohttp==3.7.3; python_version >= '3.6'
async-timeout==3.0.1; python_full_version >= '3.5.3'
asyncpg==0.21.0; python_full_version >= '3.5.0'
attrs==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
babel==2.9.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
dawg-python==0.7.2
docopt==0.6.2
fake-useragent==0.1.11
gino==1.0.1
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
lxml==4.6.2
multidict==5.1.0; python_version >= '3.6'
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy2==0.9.1
python-docx==0.8.10
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
soupsieve==2.1; python_version >= '3.0'
sqlalchemy==1.3.22; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.7.4.3
urllib3==1.26.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
yarl==1.6.3; python_version >= '3.6'
